Change primary key to an autoincrement integer
