+++ drinks_touch/alembic/versions/2025_07_08_2045-858dadbc3869_add_sales_table.py
@@ -1,0 +1,41 @@
+"""Add sales table
+
+Revision ID: 858dadbc3869
+Revises: 95beb17aebef
+Create Date: 2025-07-08 20:45:16.155503
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "858dadbc3869"
+down_revision: Union[str, None] = "95beb17aebef"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "sales",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column(
+            "date", sa.Date(), server_default=sa.text("CURRENT_DATE"), nullable=False
+        ),
+        sa.Column("ean", sa.Text(), nullable=False),
+        sa.PrimaryKeyConstraint("id", name=op.f("pk_sales")),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("sales")
+    # ### end Alembic commands ###
