+++ drinks_touch/database/models/__init__.py
@@ -5,4 +5,4 @@ from .scan_event import ScanEvent
 from .tetris import TetrisGame, TetrisPlayer
 from .tx import Tx
 from .appsettings import AppSettings
-from .sales import Sales
+from .sale import Sale
