+++ drinks_touch/database/models/sale.py
@@ -1,13 +1,11 @@
-import uuid
-
-from sqlalchemy import Column, UUID, Text, Date
+from sqlalchemy import Column, Text, Date, Integer
 from sqlalchemy.sql import func
 
 from database.storage import Base
 
 
-class Sales(Base):
+class Sale(Base):
     __tablename__ = "sales"
-    id = Column(UUID(), primary_key=True, default=uuid.uuid4)
+    id = Column(Integer, primary_key=True)
     date = Column(Date(), nullable=False, server_default=func.current_date())
     ean = Column(Text(), nullable=False)
